Updated gui appearance
- gave each item stack more information
- standardized gui item stack layout
- added option to reload configs from disk
- added option to go back in config and copy action
- adjusted config gui layout with the rest of the gui
- added option to abort copy action

Took 4 hours 16 minutes

+++ ore-control/src/main/java/de/derfrzocker/ore/control/command/set/SetValueCommand.java
@@ -125,7 +125,7 @@ public class SetValueCommand implements TabExecutor { //TODO "merge" set and set
                 return;
             }
 
-            final double value2 = percents ? (service.getDefaultValue(ore, setting) * (value / 100)) : (double) value;
+            final double value2 = percents ? (service.getDefaultValue(ore, setting) * (value / 100)) : value;
 
             if (OreControlUtil.isUnSafe(setting, value2)) {
                 if (oreControlValues.getConfigValues().isSafeMode()) {
